src/main/java/com/legadi/cli/jurl/model/http/HTTPRequestFileEntry.java
package com.legadi.cli.jurl.model.http;

public class HTTPRequestFileEntry {

    private String name;
    private String path;
    private String field;
    private String mineType;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getPath() {
        return path;
    }

    public void setPath(String path) {
        this.path = path;
    }

    public String getField() {
        return field;
    }

    public void setField(String field) {
        this.field = field;
    }

    public String getMineType() {
        return mineType;
    }

    public void setMineType(String mineType) {
        this.mineType = mineType;
    }
}
